Use AD-specific LDAPServer instance for AD tests only
